# Consolidate and optimize TextInfo.ChangeCase (string/char.ToUpper/Lower)