… src/mscorlib/shared/System/Globalization/TextInfo.Unix.cs