@danmosemsft 2018-04-03
This affects char.ToLower/Upper aka TextInfo.ToLower(char) but Perf_Char is almost empty. Do we have enough coverage of the char overloads and if not would you consider adding enough to make sure you didn't regress them?