@stephentoub 2018-04-03
> Do we have enough coverage of the char overloads and if not would you consider adding enough to make sure you didn't regress them?

Added at https://github.com/dotnet/corefx/pull/28765.  The results are the same before/after this PR.