@stephentoub 2018-04-03
@dotnet-bot test OSX10.12 x64 Checked Innerloop Build and Test please
@dotnet/dnceng, FYI:
https://ci.dot.net/job/dotnet_coreclr/job/master/job/x64_checked_osx10.12_innerloop_flow_prtest/2534/
```
FATAL: java.nio.channels.ClosedChannelException
java.nio.channels.ClosedChannelException
Also:   hudson.remoting.Channel$CallSiteStackTrace: Remote call to JNLP4-connect connection from 131.107.58.4/131.107.58.4:63995
```

@dotnet-bot test Ubuntu arm Cross Checked Innerloop Build and Test please
@dotnet/dnceng, FYI:
https://ci.dot.net/job/dotnet_coreclr/job/master/job/arm_cross_checked_ubuntu_innerloop_flow_prtest/269/
```
FATAL: command execution failed
java.nio.channels.ClosedChannelException
	at org.jenkinsci.remoting.protocol.impl.ChannelApplicationLayer.onReadClosed(ChannelApplicationLayer.java:208)
```