@tarekgh 2018-04-03
Thanks, @stephentoub for optimizing this. I think the gain will be very valuable.